Add block formatter for R6 classes
It writes out the entries for the main
part (i.e. the part that does not belong to
methods) first, e.g. description, details,
main sections, etc. All tags that are not
`@details`, `@param`, `@return` and `@examples`
automatically got into the main part. These
four tags got into the 'Methods' section, if
they appear in the inline comments, i.e. the
comments within the class definition.

The 'Methods' section has the two parts. First
there is a list of all public methods, with
in-page links. Then there is one subsection
for each method. These appear in the same order
as in the source file.

For each method we have:
- An HTML anchor, so the method can be linked
  via `[link text](#method-mymethod)`
  (from the same page), and with
  `[link text][myclass#method-mymethod]` from other
  pages.
- A "usage" part (not a subsection), with
  the formals of the method.
- Details of the method, the `@details` tags for the
  method.
- An 'Arguments' subsection if the method takes
  arguments.
- A 'Returns' subsection if the method had a
  `@return` tag. (Max one is allowed.)
- An 'Examples' subsection if the method had
  any `@examples` tags.