Leave out @param from an R6 block, but inherit them into methods
The prefix (i.e. _not_ inline) docs may contain @param tags,
that the methods automatically inherit, if needed. The methods
can also override these arguments, in which case they won't be
inherited. These @param tags will be left out from the man page
of the class otherwise.

If you want to document a function on the same page as the class,
e.g. a factory function, then include its arguments in a
separate block, the block of the function.

This is experimental, and might still change. One drawback is that
we cannot inherit currently from or into the method docs, but
we can probably fix this later.